remove pete files
