--- Users/peterallen/Projects/gradio/js/core/src/gradio_helper.d.ts
@@ -1,4 +1,0 @@
-export { Gradio } from "@gradio/utils";
-export type I18nFormatter = typeof formatter;
-export declare function formatter(value: string | null | undefined): string;
-export declare const reactive_formatter: import("svelte/store").Readable<typeof formatter>;
